Anki-TTS-PY/requirements.txt
customtkinter
edge-tts
pyperclip
pygame
pynput
pywin32
pystray
Pillow